requirements.txt
certifi==2026.2.25
chardet==7.4.0.post1
charset-normalizer==3.4.6
idna==3.11
numpy==2.4.3
pandas==3.0.1
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.0
six==1.17.0
urllib3==2.6.3
